requirements.txt
lightning
matplotlib
torchaudio
torchinfo
torchvision
av
jsonargparse[signatures]>=4.18.0
torchdata
soundfile
pesq
tensorboard
torch-tb-profiler